Match signedness of printf format with signedness of argument. Note, we cannot use the gcc option -Wformat-signedness since currently the numpy include file has a mismatch.

+++ tools/mserver/mserver5.c
@@ -444,7 +444,7 @@ main(int argc, char **av)
 			usage(prog, strcmp(av[optind - 1], "-?") == 0 || strcmp(av[optind - 1], "--help") == 0 ? 0 : -1);
 		default:
 			fprintf(stderr, "ERROR: getopt returned character "
-							"code '%c' 0%o\n", c, c);
+				"code '%c' 0%o\n", c, (uint8_t) c);
 			usage(prog, -1);
 		}
 	}
